launchpad-ledger: Tidy auxiliary files
