--- packages/launchpad-ledger/.nycrc.yml
@@ -1,1 +1,0 @@
-../../.nycrc.yml
