+++ packages/launchpad-ledger/typedoc.js
@@ -3,7 +3,7 @@ const packageJson = require("./package.json");
 module.exports = {
   inputFiles: ["./src"],
   out: "docs",
-  exclude: "**/*.spec.ts",
+  exclude: ["**/*.spec.ts", "./src/demo"],
   name: `${packageJson.name} Documentation`,
   readme: "README.md",
   mode: "file",
